batman-adv: Still flood multicast packets we are not a receiver of
We may only optimize the multicast packet flow, if an mcast_mode has been
activated and if we are a multicast receiver of the same group. Otherwise
flood the multicast packet without optimizations.

This allows us to still flood multicast packets of protocols where it is
not easily possible for a multicast sender to be a multicast receiver of
the same group instead of dropping them (for instance IPv6 NDP).

This commit therefore also makes IPv6 usable again, if the
proact_tracking multicast mode has been activated.