batman-adv: Adding sysfs parameter for ndp interval
This parameter can be set individually on each interface and allows the
configuration of the ndp interval for the link quality measurements
during runtime. Usually it is desirable to set it to a higher (= slower) value
on interfaces which have a more static characteristic (e.g. wired
interfaces) or very dense neighbourhoods to reduce overhead.

Signed-off-by: Linus Lüssing <linus.luessing@ascom.ch>