batman-adv: Add bridge multicast snooping support for MCA attachements
So far the batman-adv multicast optimization was not able to trigger
for bridged-in hosts as IGMP/MLD snooping is needed for that. Such hosts
are not receiving these multicast data packets at all.

The kernel's bridge code has a multicast snooping feature since 2.6.34
for IPv4 and since 2.6.35 for IPv6. With this commit, we are making use
of that feature and utilize the bridge's knowledge about other multicast
listeners behind the bridge and are adding these to our own OGMs as
additional MCA entries.

This patch also adds an according debugfs file.