batman-adv: Reduce local RQ in case of outdated information
This commit introduces the reduction of local RQ information
depending on the time we last heard a message of the according neighbor.
Therefore an originator's ndp interval is published in its own
NDP packets. This fixes a conceptual "race condition", in scenarios
where we are experiencing a rapid decrease of the link quality. In such
cases, the local RQ/TQ values stale at high values and occasional OGMs
could still pick them up.

With this new information the local RQ window gets shifted according to
the number of packets we expect to have missed. This check + shifting is
done during every purge interval, which currently is once per second.

Please note, that OGMs are actually interested in the local TQ
information instead, not the local RQ. However, due to the asymmetric
penalty, a local RQ degrading also has an impact on the global TQ value
of an OGM.

Signed-off-by: Linus Lüssing <linus.luessing@ascom.ch>