batman-adv: Add reactive multicast flow table management
This commit adds the possibility to count and keep track of the own
multicast flow coming in from the soft interface. This allows us to only
build up the forwarding infrastructure later if a certain threshold of
incoming multicast packets of a certain group (here multicast mac
address) is reached. Only the IPv4 non-link-local and IPv6 transient
multicast flow is being looked at and optimized.

This feature also includes a grace period timer to ensure that the
optimized multicast infrastructure has been build up before trying to
use it.

It also includes the according sysfs and debugfs files.