batman-adv: Send own BAT_MCAST packets in multicast group aware mode
This patch adds the capability to encapsulate and send a node's own
multicast data packets. Based on the previously established multicast
forwarding table, the sender can decide whether it actually has to send
the multicast data to one or more of its interfaces or not.

Furthermore, the sending procedure also decides whether to broadcast or
unicast a multicast data packet to its next-hops, depending on the
mcast_fanout, which can be configured via sysfs (default: < 16 next hops
on an interface, send separate unicast packets).