batman-adv: Code and transmit packets if possible.
Before adding forward-skbs to the coding buffer, the buffer is searched
for a potential coding opportunity. If one is found, the two packets are
network coded and transmitted right away. If not, the forward-skb is
added to the buffer.

Network coded packets are transmitted with information about the two
receivers and the two coded packets. The first receiver is given by the
MAC header, while the second is given in the payload/bat-header. The
second receiver uses promiscuous mode to receive the packet and check
the second destination.

---
v3:
 * Fix header-diff in batadv_nc_code_packets()
v4:
 * Add missing kernel doc entry for enum batadv_counters