batman-adv: Save overheard and tx packets for decoding.
To be able to decode a network coded packet, a node must already know
one of the two coded packets. This is done by buffering skbs before
transmission and buffering packets sniffed with promiscuous mode from
other hosts.

Packets are kept in a buffer similar to the one with forward-skbs: A
hash table, where each entry, which corresponds to a src-dst pair, has a
linked list packets.

---
v3:
 * Use new return values from batadv_check_unicast_packet()
v4:
 * Added missing kernel doc entry for enum batadv_counters