batman-adv: Receive coded packets and decode them.
When receiving a network coded packet, the decoding buffer is searched
for a packet to use for decoding. The source, destination, and crc32 from
the coded packet is used to identify the wanted packet. The decoded
packet is passed to the usual unicast receiver function, as had it never
been network coded.

---
v3:
 * Free nc_packet in batadv_nc_recv_coded_packet() to avoid mem-leaks
v4:
 * Add missing kernel doc entry for enum batadv_counters