batman-adv: Add blocking of one hop OGM messages
For testing purposes, forcing specific paths in a network where all
nodes are within reach, can be useful. This patch allows the user to
enter addresses from which direct OGMs should be ignored or not dropped.

The blocking of OGM is controlled through a file in debugfs:
/sys/kernel/debug/batman-adv/bat0/block_ogm.

The list of currently handled addresses is exported through that file.
An addresses is added by echoing it, together with an action, into the
file, e.g.:
echo de:ad:be:ef:ca:fe drop  > block_ogm
echo de:ad:be:ef:ca:fe allow > block_ogm
echo de:ad:be:ef:ca:fe del   > block_ogm

The drop action leads to dropping all OGMs received directly from this
address. The allow action leads to dropping all OGMs *not* from this
(or other allowed) addresses. The del action deletes the address from
the blocking list.

Addresses and actions are stored in a doubly linked list. When new
addresses are added, the list of originators is searched, and an
possible existing originator is updated. When new originators are created,
the list of addresses is searched and "orig_node->block_action" is set
accordingly.

The filtering is performed by checking the value of member
"orig_node->block_action" together with a count of addresses with the
"allow" action. The orig_node struct is found using the Ethernet source of
the OGM message in question. If "orig_node->block_action" is equal 2. If
the count of addresses with the "allow" action is non-zero, and
"orig_node->block_action" is not 3, the OGM is dropped.

v2: This is almost a complete rewrite of the previous patch to allow
adding addresses of unknown originators and the allow-only feature.