batman-adv: bandwidth meter implementation
The bandwith meter module is a simple, kernel-space replacement for bandwith
measurements tool like iperf and netper. It is intended to approximate TCP
behaviour.

It is invoked through batctl: the protocol is connection oriented, with
cumulative acknowledgment and sliding window. Sender keeps a timeout, which is
checked by a worker function regoularly invoked through the workqueue mechanism.
If the timeout is expired at the time worker is executed, the whole window is
re-transmitted.

Sequence number in packet header is 16 bit and is wrap around safe.
BATADV_BW_FIRST_SEQ is the sequence number of the first packet. It is set to
65530 to generate a wrap-around. *Both sides must agree on this value*.

The receiver maintains a bitmap of BATADV_BW_WINDOW_SIZE to account for received
packets not in order, so to avoid that they are unecessarly resent. Every time a
packet is received, if it is either in order or not, an ack with the seqno
equal to the last in order packet is sent back.

The test *can* be interrupted by batctl. A receiver side timeout avoids unlimited
waitings for sender packets: after one second of inactivity, the receiver abort
the ongoing test. The sender aborts the test after having resent
BATADV_BW_MAX_RETRY the same window.

When the test is over, the results are returned to batctl through a call to the
function batadv_socket_receive_packet(), before freeing struct batadv_bw_vars.
This function only accepts struct batadv_icmp_packet_rr, so that structure is
used and a cast is done to struct batadv_bw_result. The function basically sends
a particular icmp packet through the local socket. In case any error occured
during the test, this is also reported to batctl.

Signed-off-by: Edo Monticelli <montik@autistici.org>