batman-adv: bandwidth meter implementation
The bandwith meter module is a simple, kernel-space replacement for bandwith
measurements tool like iperf and netper. It is intended to approximate TCP
behaviour.

It is invoked through batctl: the protocol is connection oriented, with
cumulative acknowledgment and sliding window. Sender keeps a timeout, which is
checked by a worker function regoularly invoked through the workqueue mechanism.
If the timeout is expired at the time worker is executed, the whole window is
re-transmitted.

Struct bw_vars, one for each connection, are collected into a linked list. Thus
access to them is protected with a spinlock.

Another spinlock prevents more than one instance of multiple_send to run
concurrently. Ack code is mutually exclusive (two ack handler cannot run at the
same time for differnent acks), but they are not mutually exclusive with the
function responsible for sending the window (multiple_send), that is to say the
ack handler shifts the window ``while" multiple_send is running. Maybe some
atomic variable (window_first) is needed.

When the test is over, the results are returned to batctl through a call to the
function batadv_socket_receive_packet(), before freeing struct batadv_bw_vars.
This function only accepts struct batadv_icmp_packet_rr, so that structure is
used and a cast is done to struct batadv_bw_result.

The test interruptable with CTRL-C. A receiver side timeout avoids unlimited
waitings for sender packets: after one second of inactivity, the receiver abort
the ongoing test. The sender aborts the test after having resent
BATADV_BW_MAX_RETRY the same window.

Sequence number in packet header is 16 bit and is wrap around safe. Two 32 bit
counters, window_first and total_to_send are kept locally in the struct bw_vars.
On packet or acks arrival batadv_seq_before()/after() are used to check seqno
against 16bit version of these counters.  BATADV_BW_FIRST_SEQ is the sequence
number of the first packet. It is set to 65530 to generate a wrap-around. *Both
sides must agree on this value*

Signed-off-by: Edo Monticelli <montik@autistici.org>