bw_meter and relative struct first skeletron
File bw_meter.c and bw_meter.h are added

In them functions to start the test is defined. Function that is invoked when an
icmp packet of type BW_METER and BW_ACK also is defined here.

types.h and packet.h are modified to introduce the new icmp_packet_bw format and
the data structure bw_meter_vars, added to bat_priv.

The other files included just contain minor modifications, to call bw_meter
functions when an event that need one of them happen (i.e. the reception of an
icmp_packet_bw).

Signed-off-by: montik <montik@autistici.org>