batman-adv: introduces locking and bw_vars list
Struct bw_vars are collected into a linked list. Thus access to them is
protected with a spinlock.

Two more spinlocks have been introduced, one for the ack and one for the send.
This is due to prevent multiple instance of send_remaining_window to run
concurrently, messing up things and accessing the same variables.

Ack code is mutually exclusive (two ack handler cannot run at the same time for
differnent acks), but they are not mutually exclusive with
send_remaining_window, that is to say the ack handler shifts the window ``while"
send_remaining_window is running. Maybe some atomic variable (window_first) is
needed to avoid problem on this.

Test works, but it's not exhaustive.

Signed-off-by: Edo Monticelli <montik@autistici.org>