Atomic code reduced
The portions of code that needs lock have been reduced, hopefully obtaining a
performance improvement.
Two new spinlocks have been introduced, one for the ack and one for the send.
This has the advantage to prevent multiple instance of send_remaining_window to
be run concurrently, messing up things and accessing the same variable.

Ack code is mutually exclusive (two ack handler cannot run at the same time for
differnent acks), but they are not mutually exclusive with
send_remaining_window, that is to say the ack handler shifts the window ``while"
send_remaining_window is running. Maybe some atomic variable (window_first) is
needed to avoid problem on this.

Test works, but it's not exhaustive.

Signed-off-by: montik <montik@autistici.org>