batman-adv: integration with batctl
When the test is over, the results are returned to batctl through a call to the
function batadv_socket_receive_packet(), before freeing struct batadv_bw_vars.

The function only accepts struct batadv_icmp_packet_rr, so that structure is used
and a cast is done to struct batadv_bw_result.

Signed-off-by: Edo Monticelli <montik@autistici.org>