batman-adv: bitmap on receiver side
The receiver maintains a bitmap of BATADV_BW_WINDOW_SIZE to account for received
packets not in order. Every time a packet is received an ack is sent back,
if it is either in order or not, with the seqno equals to the last in order
packet.

Still some problems to fix.

Signed-off-by: Edo Monticelli <montik@autistici.org>