batman-adv: Tracker Packet compat code
If the neighbor a tracker packet is supposed to be forwarded to did not
signal any multicast optimization support, then the tracker packet for
this neighbor will be split into multiple tracker packets encapsulated
into a batman-adv unicast header for any destination of this tracker
packet.

A multicast aware node receiving such an encapsulated packet will remove
the unicast header again (if the next hop is a multicast aware node).