batman-adv: Fix a potential bcast/ogm queue purging race condition (2)
- Avoid rearmed bcast packet being missed by cancel_delayed_work_sync().

When a scheduled broadcast packet sending work executes and removes
itself from the broadcast queue, it becomes invisible for the purging
methods until it re-arms.

This means a to be re-armed broadcast work might slip through a
cancel_delayed_work_sync() between the removal from and readdition to
the broadcast queue.

To avoid this we perform the removal and readdition in one atomic step.

Signed-off-by: Linus Lüssing <linus.luessing@web.de>