batman-adv: Don't re-broadcast packets on non-wifi devices
It is not necessary to re-broadcast a received broadcast packet on the
rx-device when it is a standard ethernet device. The link medium already
takes care of transporting it to all participants in the broadcast domain.

The re-broadcast on other devices is still necessary to allow the broadcast
packet to be received by devices which are using a different link medium.

Signed-off-by: Sven Eckelmann <sven@narfation.org>